# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface Mailbox {
  id: number;
  address: string;
  name?: string | null;
  createdAt: string;
  expiresAt?: string | null;
  lastActivity?: string | null;
  messageCount: number;
  unreadCount: number;
}

export interface ListMailboxesResponse {
  mailboxes: Mailbox[];
  total: number;
}

export interface CreateMailboxBody {
  /** The part before @domain. If omitted, a random one is generated. */
  localPart?: string | null;
  /** Optional display name */
  name?: string | null;
}

export interface MessageSummary {
  id: number;
  mailboxAddress: string;
  fromAddress: string;
  fromName?: string | null;
  subject: string;
  receivedAt: string;
  isRead: boolean;
  hasHtml: boolean;
}

export interface MessageDetail {
  id: number;
  mailboxAddress: string;
  fromAddress: string;
  fromName?: string | null;
  subject: string;
  bodyText?: string | null;
  bodyHtml?: string | null;
  receivedAt: string;
  isRead: boolean;
}

export interface ListMessagesResponse {
  messages: MessageSummary[];
  total: number;
  unread: number;
}

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  createdAt: string;
  lastUsedAt?: string | null;
}

export interface ApiKeyWithSecret {
  id: number;
  name: string;
  keyPrefix: string;
  /** Full API key - only shown once on creation */
  key: string;
  createdAt: string;
}

export interface ListApiKeysResponse {
  keys: ApiKey[];
}

export interface CreateApiKeyBody {
  name: string;
}

export interface InboundEmailBody {
  /** Recipient email address */
  to: string;
  /** Sender email address */
  from: string;
  /** Sender display name */
  fromName?: string | null;
  subject: string;
  bodyText?: string | null;
  bodyHtml?: string | null;
  /** Webhook secret for verification */
  secret: string;
}

export interface InboundEmailResponse {
  success: boolean;
  messageId?: number | null;
}

export interface StatsResponse {
  totalMailboxes: number;
  totalMessages: number;
  totalUnread: number;
  domain: string;
}