# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * Returns all created mailboxes
 * @summary List all mailboxes
 */
export const ListMailboxesResponse = zod.object({
  mailboxes: zod.array(
    zod.object({
      id: zod.number(),
      address: zod.string(),
      name: zod.string().nullish(),
      createdAt: zod.coerce.date(),
      expiresAt: zod.coerce.date().nullish(),
      lastActivity: zod.coerce.date().nullish(),
      messageCount: zod.number(),
      unreadCount: zod.number(),
    }),
  ),
  total: zod.number(),
});

/**
 * Create a new mailbox with a custom or randomly generated address
 * @summary Create a mailbox
 */
export const CreateMailboxBody = zod.object({
  localPart: zod
    .string()
    .nullish()
    .describe(
      "The part before @domain. If omitted, a random one is generated.",
    ),
  name: zod.string().nullish().describe("Optional display name"),
});

/**
 * @summary Get mailbox info
 */
export const GetMailboxParams = zod.object({
  address: zod.coerce.string(),
});

export const GetMailboxResponse = zod.object({
  id: zod.number(),
  address: zod.string(),
  name: zod.string().nullish(),
  createdAt: zod.coerce.date(),
  expiresAt: zod.coerce.date().nullish(),
  lastActivity: zod.coerce.date().nullish(),
  messageCount: zod.number(),
  unreadCount: zod.number(),
});

/**
 * @summary Delete a mailbox
 */
export const DeleteMailboxParams = zod.object({
  address: zod.coerce.string(),
});

/**
 * @summary List messages in a mailbox
 */
export const ListMessagesParams = zod.object({
  address: zod.coerce.string(),
});

export const ListMessagesResponse = zod.object({
  messages: zod.array(
    zod.object({
      id: zod.number(),
      mailboxAddress: zod.string(),
      fromAddress: zod.string(),
      fromName: zod.string().nullish(),
      subject: zod.string(),
      receivedAt: zod.coerce.date(),
      isRead: zod.boolean(),
      hasHtml: zod.boolean(),
    }),
  ),
  total: zod.number(),
  unread: zod.number(),
});

/**
 * @summary Get a specific message
 */
export const GetMessageParams = zod.object({
  address: zod.coerce.string(),
  id: zod.coerce.number(),
});

export const GetMessageResponse = zod.object({
  id: zod.number(),
  mailboxAddress: zod.string(),
  fromAddress: zod.string(),
  fromName: zod.string().nullish(),
  subject: zod.string(),
  bodyText: zod.string().nullish(),
  bodyHtml: zod.string().nullish(),
  receivedAt: zod.coerce.date(),
  isRead: zod.boolean(),
});

/**
 * @summary Delete a message
 */
export const DeleteMessageParams = zod.object({
  address: zod.coerce.string(),
  id: zod.coerce.number(),
});

/**
 * @summary Mark message as read
 */
export const MarkMessageReadParams = zod.object({
  address: zod.coerce.string(),
  id: zod.coerce.number(),
});

export const MarkMessageReadResponse = zod.object({
  id: zod.number(),
  mailboxAddress: zod.string(),
  fromAddress: zod.string(),
  fromName: zod.string().nullish(),
  subject: zod.string(),
  receivedAt: zod.coerce.date(),
  isRead: zod.boolean(),
  hasHtml: zod.boolean(),
});

/**
 * @summary List API keys
 */
export const ListApiKeysResponse = zod.object({
  keys: zod.array(
    zod.object({
      id: zod.number(),
      name: zod.string(),
      keyPrefix: zod.string(),
      createdAt: zod.coerce.date(),
      lastUsedAt: zod.coerce.date().nullish(),
    }),
  ),
});

/**
 * @summary Create an API key
 */
export const CreateApiKeyBody = zod.object({
  name: zod.string(),
});

/**
 * @summary Delete an API key
 */
export const DeleteApiKeyParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Receive inbound email from Cloudflare Worker
 */
export const InboundEmailBody = zod.object({
  to: zod.string().describe("Recipient email address"),
  from: zod.string().describe("Sender email address"),
  fromName: zod.string().nullish().describe("Sender display name"),
  subject: zod.string(),
  bodyText: zod.string().nullish(),
  bodyHtml: zod.string().nullish(),
  secret: zod.string().describe("Webhook secret for verification"),
});

export const InboundEmailResponse = zod.object({
  success: zod.boolean(),
  messageId: zod.number().nullish(),
});

/**
 * @summary Get service stats
 */
export const GetStatsResponse = zod.object({
  totalMailboxes: zod.number(),
  totalMessages: zod.number(),
  totalUnread: zod.number(),
  domain: zod.string(),
});