# lib/api-zod/src/generated/types/apiKey.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface ApiKey {
  id: number;
  name: string;
  keyPrefix: string;
  createdAt: Date;
  lastUsedAt?: Date | null;
}