# lib/api-zod/src/generated/types/createApiKeyBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface CreateApiKeyBody {
  name: string;
}