# lib/api-zod/src/generated/types/createMailboxBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMailboxBody {
  /** The part before @domain. If omitted, a random one is generated. */
  localPart?: string | null;
  /** Optional display name */
  name?: string | null;
}