# lib/api-zod/src/generated/types/inboundEmailBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface InboundEmailBody {
  /** Recipient email address */
  to: string;
  /** Sender email address */
  from: string;
  /** Sender display name */
  fromName?: string | null;
  subject: string;
  bodyText?: string | null;
  bodyHtml?: string | null;
  /** Webhook secret for verification */
  secret: string;
}