# lib/api-zod/src/generated/types/listApiKeysResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */
import type { ApiKey } from "./apiKey";

export interface ListApiKeysResponse {
  keys: ApiKey[];
}