# lib/api-zod/src/generated/types/listMessagesResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */
import type { MessageSummary } from "./messageSummary";

export interface ListMessagesResponse {
  messages: MessageSummary[];
  total: number;
  unread: number;
}