# lib/api-zod/src/generated/types/mailbox.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface Mailbox {
  id: number;
  address: string;
  name?: string | null;
  createdAt: Date;
  expiresAt?: Date | null;
  lastActivity?: Date | null;
  messageCount: number;
  unreadCount: number;
}