# lib/api-zod/src/generated/types/messageDetail.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API - Temporary email service using custom domain
 * OpenAPI spec version: 0.1.0
 */

export interface MessageDetail {
  id: number;
  mailboxAddress: string;
  fromAddress: string;
  fromName?: string | null;
  subject: string;
  bodyText?: string | null;
  bodyHtml?: string | null;
  receivedAt: Date;
  isRead: boolean;
}